RND203-337: Align headers with RND203-327
